requirements.txt
crewai==0.121.1
crewai-tools==0.46.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pandas>=2.0.0
cryptography>=41.0.0  # Required for PyMySQL
langchain-openai>=0.0.3
python-dotenv>=1.0.0 